able to log motor data, with velocity control, but the measured values are rounded for some reason

+++ CCU/src/i2c_master.h
@@ -12,7 +12,7 @@
 
 struct MUData
 {
-    uint8_t setpoint_recv;
+    float setpoint_recv;
     float value_recv;
     float current_recv;
 };
